# Raw bit generator sampling