--- a/starsim/disease.py
+++ b/starsim/disease.py
@@ -152,8 +152,8 @@ class Infection(Disease):
         )
 
         # Define random number generators for make_new_cases
-        self.rng_target = ss.randint(name='target')
-        self.rng_source = ss.randint(name='source')
+        self.rng_target = ss.rand_uint64(name='target')
+        self.rng_source = ss.rand_uint64(name='source')
         return
     
     def init_pre(self, sim):