--- a/starsim/distributions.py
+++ b/starsim/distributions.py
@@ -792,6 +792,8 @@ class randint(Dist):
 class rand_uint64(Dist):
     def make_rvs(self):
         # The bit generator random_raw function is specified as returning uint64
+        if ss.options._centralized:
+            return self.rng._bit_generator.random_raw(self._size)
         return self.bitgen.random_raw(self._size)