--- a/tests/benchmark.json
+++ b/tests/benchmark.json
@@ -1,12 +1,12 @@
 {
   "time": {
-    "initialize": 0.018,
-    "run": 1.264
+    "initialize": 0.015,
+    "run": 0.92
   },
   "parameters": {
     "n_agents": 10000,
     "n_years": 20,
     "dt": 0.2
   },
-  "cpu_performance": 0.9766497596430002
+  "cpu_performance": 0.6604880837973948
 }