--- a/tests/test_randomness.py
+++ b/tests/test_randomness.py
@@ -291,8 +291,8 @@ def test_combine_rands(do_plot=False):
     atol = 1e-3
     target = 0.5
     np.random.seed(2)
-    a = np.random.randint(np.iinfo(np.int64).max, size=n)
-    b = np.random.randint(np.iinfo(np.int64).max, size=n)
+    a = np.random.randint(np.iinfo(np.uint64).max, size=n, dtype=np.uint64)
+    b = np.random.randint(np.iinfo(np.uint64).max, size=n, dtype=np.uint64)
     c = ss.combine_rands(a, b)
     if do_plot:
         pl.figure()