### Description

The issue observed in #568 appears to be because `rng.random()` (to get a random float) is faster than `rng.integers()`. It's been suggested that this is because `random()` is a pure C call while `integers()` has some Python logic as well. In the test case noted by @cliffckerr the simple model runtime is dominated by the transmission portion of the calculation and this difference is big enough to provide a noticeable difference in runtime.

@daniel-klein had suggested a while ago that maybe we could just use the numpy bit generator directly to get the random integers we need. I've implemented this in a new `Dist`, `rand_uint64`,  that calls `bitgen.random_raw` to get the `uint64` array of random integers that is then passed to `combine_rands`. We're already asking users who want to use `combine_rands` to create `Dist` instances with integers even though the final random number is going to be a float, so they have to view the setup as an integrated package (with integer `Dists` and `combine_rands`) so I don't think it's too onerous (and in fact, maybe a bit simpler/clearer) to just say `rand_uint64` provides inputs for use with `combine_rands` (could potentially give it a better name).

Performance is a little bit tricky because the current `main` branch has had other optimizations in the meantime. Running @cliffckerr 's test script from #568 gives

- Old `main` (`43249e3`): 733 ms ± 3.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- `combine_rands` (`4689698`): **10% slower** (810 ms ± 6.61 ms per loop (mean ± std. dev. of 7 runs, 1 loop each))
- `main` (`4e0bcfc`): **12% faster** (641 ms ± 15.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each))
- `test_raw_rng`: **20% faster** (582 ms ± 4.67 ms per loop (mean ± std. dev. of 7 runs, 1 loop each))

So overall `test_raw_rng` has higher performance than the old `main` branch where `combine_rands` had worse performance. The impact of this issue depends on what proportion of time in a model is spent calculating transmission, so for a complex model like STIsim, it makes almost no difference to the runtime because so much time is spent on other parts of the model). 

Not sure about whether it's worth the added complexity of a separate integer generator just for this purpose, but on the other hand, the performance _is_ better and it solves #568, so opening this PR for comment/review

### Checklist
- [ ] Code commented & docstrings added
- [ ] New tests were needed and have been added
- [ ] A new version number was needed & changelog has been updated
- [ ] A new PyPI version needs to be released