@daniel-klein 2024-07-11
Thanks @RomeshA, this looks amazing. I saw the `random_raw` as well, but at that point was convinced that we needed signed integers. In response to your comment on the paper, I've done a bit more archeology and no longer see an advantage either way.