@RomeshA 2024-07-12
@daniel-klein Weird! It looked fine with Cliff's 2D histogram tests, and a reduced run of `explore_tx_correlation.py` also seemed OK - what was the script that produced those biased results?