@daniel-klein 2024-07-12
@RomeshA https://github.com/starsimhub/crn_paper/blob/main/explore_erdosrenyinet.py