@daniel-klein 2024-07-12
I'm wondering if it's related to this warning in the [random_raw](https://numpy.org/doc/stable/reference/random/bit_generators/generated/numpy.random.BitGenerator.random_raw.html) documentation:
```
This method directly exposes the raw underlying pseudo-random number generator. All values are returned as unsigned 64-bit values irrespective of the number of bits produced by the PRNG.
```

For the failing test, I'm using the default Mersenne Twister.