@daniel-klein 2024-07-12
Ah yes, that's probably it. From the [Mersenne Twister (MT19937)](https://numpy.org/doc/stable/reference/random/bit_generators/mt19937.html) docs:
```
The [MT19937](https://numpy.org/doc/stable/reference/random/bit_generators/mt19937.html#numpy.random.MT19937) state vector consists of a 624-element array of 32-bit unsigned integers plus a single integer value between 0 and 624 that indexes the current position within the main array.
```