@daniel-klein 2024-07-12
"The [PCG64DXSM](https://numpy.org/doc/stable/reference/random/bit_generators/pcg64dxsm.html#numpy.random.PCG64DXSM) state vector consists of 2 unsigned 128-bit values, which are represented externally as Python ints."

Not sure how well this will work, might just be inefficient to return as `uint64`, but running test now.