@daniel-klein 2024-07-12
Yay, all good with PCG64DXSM. I can fix `_centralized` with something like the following
```python
np.random.mtrand._rand.randint(low=0, high=np.iinfo(np.uint64).max, dtype=np.uint64, size=n)
```